ch09/makdo/pyproject.toml
[project]
name = "makdo"
version = "0.1.0"
description = "MAKDO - Multi-Agent Kubernetes DevOps System"
readme = "README.md"
authors = [
    { name = "Gigi Sayfan", email = "the.gigi@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "ai-six @ file:///Users/gigi/git/ai-six/py",
    "pyyaml",
    "python-dotenv",
]

[project.scripts]
makdo = "makdo:cli_main"

[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"
